// backend/src/shop/types/shop.types.ts
export enum ShopCategory {
  GENEL = 'genel',
  VIP = 'vip',
  LONCA = 'lonca',
  ETKINLIK = 'etkinlik',
}

export enum ProductTag {
  NEW = 'new',
  BEST = 'best',
  LIMITED = 'limited',
  HOT = 'hot',
}

export enum Currency {
  GEM = 'gem',
  GOLD = 'gold',
}

export enum PurchaseStatus {
  PENDING = 'pending',
  COMPLETED = 'completed',
  FAILED = 'failed',
}